Added white dwarves and brown dwarves, bug fixes
Added white and brown dwarves, rewrote some code to handle "special stars" better. Changed luminosity and radius calculation. Fixed errors with asteroid belt and chunks. Fixed some errors related to data being handled badly. Other bug fixes and typo corrections were done as well.

+++ Planet.h
@@ -24,7 +24,7 @@ public:
     double perihelion;
     //Planet Values
     enum class planet_type { asteroid_belt, terrestrial, chunk, gas_giant, superjovian } planet_class; //0 - Asteroid Belt, 1 - Terrestrial, 2 - Chunk, 3 - Gas Giant, 4 - Superjovian
-    double radius; //in km
+    long double radius; //in km
     double density; //in earth densities
     double mass; //in earth masses
     double gravity; //in earth gravities
